[Changes] - File changes done (#19)
* file changes done

* endormanent page file removed

* fixed prettier issues

---------
