+++ src/pages/_app.jsx
@@ -1,0 +1,5 @@
+import "@/styles/global.css";
+
+export default function MyApp({ Component, pageProps }) {
+    return <Component {...pageProps} />;
+}
