+++ tsconfig.json
@@ -23,6 +23,13 @@
             "@/*": ["./src/*"]
         }
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+    "include": [
+        "next-env.d.ts",
+        "**/*.ts",
+        "**/*.tsx",
+        ".next/types/**/*.ts",
+        "src/pages/_app.js",
+        "src/pages/_app.jsx"
+    ],
     "exclude": ["node_modules"]
 }
